moved classes needed fro RPN to misc
